Update pb upload to not remove old recordings if wr

+++ src/app/common/firestore/db-run.ts
@@ -141,9 +141,11 @@ export class DbRun {
                                 const oldPbSubscription = firestoreService.getPb(previousPb.id).subscribe(oldDbPb => {
                                     oldPbSubscription.unsubscribe();
                                     if (oldDbPb) {
-                                        oldDbPb.playbackAvailable = false;
                                         oldDbPb.isCurrentPb = false;
-                                        firestoreService.deleteRecording(oldDbPb.id);
+                                        if (!oldDbPb.wasWr) {
+                                            oldDbPb.playbackAvailable = false;
+                                            firestoreService.deleteRecording(oldDbPb.id);
+                                        }
                                         firestoreService.updatePb(oldDbPb);
                                     }
                                 });
